rename operator -> installstrategies
